added react-router-dom, begining integration
