+++ package.json
@@ -11,6 +11,7 @@
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-icons": "^4.10.1",
+		"react-router-dom": "^6.14.2",
 		"react-scripts": "5.0.1",
 		"web-vitals": "^2.1.4"
 	},
