+++ src/pages/UserWorkoutPage.js
@@ -1,0 +1,12 @@
+import Navbar from "../components/Navbar";
+
+function UserWorkoutPage(props) {
+	return (
+		<div>
+			<Navbar />
+			<div>Userworkout Page</div>
+		</div>
+	);
+}
+
+export default UserWorkoutPage;
